Traslado de funciones de persistencia al módulo persistence.py
