+++ semana_17/project_finance_manager/project/gui.py
@@ -1,8 +1,7 @@
 
-import json
-import os
 import FreeSimpleGUI as sg
 from manager import FinanceManager
+from persistence import save_data, load_data
 
 
 # # # # # Start of FUNCTIONS # # # # #
@@ -94,32 +93,8 @@ def show_add_category_window(categories):
             window.close()
             return values
 
+
 # # # # # End of FUNCTIONS # # # # #
-
-# # # # # Start of PERSISTENCE # # # # # 
-
-def save_data(manager):
-    with open("semana_17/project_finance_manager/data.json", "w", encoding='utf-8') as file:
-        json.dump(manager.convert_all_to_dict(), file, indent=4)
-
-def load_data(manager):
-    if not os.path.exists("semana_17/project_finance_manager/data.json"):
-        return
-    
-    with open("semana_17/project_finance_manager/data.json", "r", encoding='utf-8') as file:
-        data = json.load(file)
-    
-    for category in data.get("categories", []):
-        manager.add_category(category)
-    
-    for m in data.get("movements", []):
-        if m["type"] == "income":
-            manager.add_income(m["title"], m["amount"], m["category"])
-        else:
-            manager.add_expense(m["title"], abs(m["amount"]), m["category"])
-
-# # # # # End of PERSISTENCE # # # # #
-
 
 def run_app():
     manager = FinanceManager()
